allow prepending with numbers
In case 2 files need to be named the same the name can be prepended by a
number which is ignored. This is the case since files on the filesystem
are now allowed to be named the same.

According to canned this is equivalent

  1.index.get.html
  index.get.html

The order is determent by the read order of the filesystem so no
guarantees are given. This change is made to allow further changes
regarding name due to windows not allowing certain characters in the
files names